Fixed math typo

+++ oaidc_analysis.py
@@ -161,7 +161,7 @@ def calc_completeness(stats_averages):
                                              completeness["collection_completeness"] +
                                              completeness["wwww_completeness"] +
                                              completeness["dpla_completeness"] +
-                                             completeness["niu_completeness"]) / float(4))
+                                             completeness["niu_completeness"]) / float(5))
     return completeness
